Reorganise Assessments context factories
